requirements.txt
fastapi
uvicorn
pydantic
requests
python-multipart
python-dotenv
gtts
openai
google-generativeai
